Move setup/teardown to helper module and simplify code in media.py
