+++ elodie/tests/media/photo_test.py
@@ -17,21 +17,11 @@ sys.path.insert(0, os.path.abspath(os.path.dirname(os.path.dirname(os.path.realp
 import helper
 from elodie.media.media import Media
 from elodie.media.photo import Photo
-from elodie.external.pyexiftool import ExifTool
-from elodie.dependencies import get_exiftool
-from elodie import constants
 
 os.environ['TZ'] = 'GMT'
 
-def setup_module():
-    exiftool_addedargs = [
-            u'-config',
-            u'"{}"'.format(constants.exiftool_config)
-        ]
-    ExifTool(executable_=get_exiftool(), addedargs=exiftool_addedargs).start()
-
-def teardown_module():
-    ExifTool().terminate
+setup_module = helper.setup_module
+teardown_module = helper.teardown_module
 
 def test_photo_extensions():
     photo = Photo()
